requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
gunicorn==21.2.0
google-api-core==2.17.1
google-auth==2.28.2
google-generativeai==0.4.0
protobuf==4.25.3
